+++ LibManSys-Backend/src/main/java/ca/mcgill/ecse321/projectgroup09/dao/BookingRepository.java
@@ -2,12 +2,18 @@ package ca.mcgill.ecse321.projectgroup09.dao;
 
 import java.util.List;
 import org.springframework.data.repository.CrudRepository;
-import ca.mcgill.ecse321.projectgroup09.models.*;
-
-public interface BookingRepository extends CrudRepository<Booking, Long> {
+import ca.mcgill.ecse321.projectgroup09.models.Member;
+import ca.mcgill.ecse321.projectgroup09.models.Librarian;
+import ca.mcgill.ecse321.projectgroup09.models.Booking;
+/**
+ * 
+ * @author Zarif Ashraf
+ *
+ */
+public interface BookingRepository extends CrudRepository<Booking, Integer> {
 	
 	List<Booking> findByMember(Member member);
 	List<Booking> findByLibrarian(Librarian librarian);
-	Booking findBookingByBookingID(Long bookingID);
+	Booking findBookingByBookingID(Integer bookingID);
 	
 }
