+++ LibManSys-Backend/src/main/java/ca/mcgill/ecse321/projectgroup09/dao/MusicAlbumRepository.java
@@ -3,9 +3,14 @@ package ca.mcgill.ecse321.projectgroup09.dao;
 import org.springframework.data.repository.CrudRepository;
 import ca.mcgill.ecse321.projectgroup09.models.MusicAlbum;
 import java.util.List;
-public interface MusicAlbumRepository extends CrudRepository<MusicAlbum, Long> {
+/**
+ * 
+ * @author Zarif Ashraf
+ *
+ */
+public interface MusicAlbumRepository extends CrudRepository<MusicAlbum, Integer> {
 	
-	MusicAlbum findMusicAlbumBylibraryItemID(Long libraryItemID);
+	MusicAlbum findMusicAlbumBylibraryItemID(Integer libraryItemID);
 	List<MusicAlbum> findMusicAlbumByGenre(String genre);
 	List<MusicAlbum> findMusicAlbumByArtist(String artist);
 }
